+++ GMI_GridComp/GmiEmission/llnl/emiss_llnl.F90
@@ -13,13 +13,6 @@
 !   Add_Emiss_Llnl
 !
 ! HISTORY
-!   - December 8, 2005 * Bigyani Das
-!     Changes are made for the aerocom run which are controlled by 2 parameters
-!     do_aerocom and do_dust_emiss to add dust and sea salt emissions in 
-!     the model's  1st level. Originally sea salt emissions 
-!     were added in the emission fields with carbon emission, 
-!     now it is added with dust (dust and sea salt emissions are now 
-!     in the same files, before sea salt emissions were with carbon).
 !=============================================================================
 
 
@@ -49,26 +42,20 @@
 !   const          : species concentration, known at zone centers
 !                    (mixing ratio)
 !   emiss          : array of emissions (kg/s)
-!   emiss_dust     : tbd
-!   emiss_aero     :
 !   pbl            :
 !-----------------------------------------------------------------------------
 
       subroutine Add_Emiss_Llnl  &
-     &  (do_aerocom, pr_surf_emiss, pr_emiss_3d, mcor, surf_emiss_out, emiss_3d_out,  &
-     &   mass, concentration, emissionArray, emiss_dust, emiss_aero_t, emiss_aero, pbl, &
+     &  (pr_surf_emiss, pr_emiss_3d, mcor, surf_emiss_out, emiss_3d_out,  &
+     &   mass, concentration, emissionArray, pbl, &
      &   gridBoxHeight, IBOC, IBBC, INOC, IFOC, IFBC, ISSLT1, ISSLT2, ISSLT3, ISSLT4, &
-     &   IFSO2, INSO2, INDMS, pr_diag, loc_proc, chem_opt, trans_opt, emiss_aero_opt, &
-     &   emiss_dust_opt, do_semiss_inchem, emiss_map, emiss_map_dust, emiss_map_aero, &
-     &   emissionSpeciesLayers, ndust, naero, nymd, mw, tdt, emiss_timpyr, num_emiss, &
+     &   IFSO2, INSO2, INDMS, pr_diag, loc_proc, chem_opt, trans_opt, &
+     &   do_semiss_inchem, emiss_map, &
+     &   emissionSpeciesLayers, nymd, mw, tdt, emiss_timpyr, num_emiss, &
      &   i1, i2, ju1, j2, k1, k2, ilo, ihi, julo, jhi, num_species, mixPBL)
 
       use GmiArrayBundlePointer_mod, only : t_GmiArrayBundle
       use GmiTimeControl_mod  , only : GmiSplitDateTime
-      use GmiSeaSaltMethod_mod, only : SourceSeaSalt
-      use GmiSeaSaltMethod_mod, only : srcEmissSeaSalt
-      use GmiDustMethod_mod   , only : SourceDust
-      use GmiDustMethod_mod   , only : srcEmissDust
 
       implicit none
 
@@ -85,30 +72,25 @@
       integer, intent(in   ) :: loc_proc
       integer, intent(in   ) :: i1, i2, ju1, j2, k1, k2, ilo, ihi, julo, jhi
       integer, intent(in   ) :: num_species, emiss_timpyr, num_emiss
-      integer, intent(in   ) :: ndust, naero, nymd
+      integer, intent(in   ) :: nymd
       real*8 , intent(in   ) :: mw(num_species)
       real*8 , intent(in   ) :: tdt
-      integer, intent(in   ) :: chem_opt, trans_opt, emiss_aero_opt, emiss_dust_opt
-      integer, intent(in   ) :: emiss_map(num_species), emiss_map_dust(num_species)
-      integer, intent(in   ) :: emiss_map_aero(num_species)
+      integer, intent(in   ) :: chem_opt, trans_opt
+      integer, intent(in   ) :: emiss_map(num_species)
       logical, intent(in   ) :: do_semiss_inchem
-      logical, intent(in   ) :: do_aerocom
       logical, intent(in   ) :: pr_surf_emiss, pr_emiss_3d
       real*8 , intent(in   ) :: mcor (i1:i2, ju1:j2)
       real*8 , intent(inout) :: surf_emiss_out(i1:i2, ju1:j2, num_species)
       real*8 , intent(inout) :: emiss_3d_out(i1:i2, ju1:j2, k1:k2, num_species)
       real*8 , intent(in   ) :: mass (i1:i2, ju1:j2, k1:k2)
 !      real*8 , intent(in   ) :: emiss(i1:i2, ju1:j2, k1:k2, num_emiss)
-      real*8 , intent(inout) :: emiss_dust(i1:i2, ju1:j2, ndust)
-      real*8 , intent(in   ) :: emiss_aero_t(i1:i2, ju1:j2, naero, emiss_timpyr)
-      real*8 , intent(out  ) :: emiss_aero(i1:i2, ju1:j2, naero)
       real*8 , intent(in   ) :: pbl (i1:i2, ju1:j2)
       real*8 , intent(in   ) :: gridBoxHeight(i1:i2, ju1:j2, k1:k2)
       type (t_GmiArrayBundle), intent(inout) :: concentration(num_species)
       type (t_GmiArrayBundle), intent(inout) :: emissionArray(num_emiss  )
       integer, intent(in   ) :: emissionSpeciesLayers(num_species)
       logical, intent(in   ) :: mixPBL    ! whether to explicitly distribute
-                                          ! aerosol emissions within the PBL
+                                          ! emissions within the PBL
 
 !     ----------------------
 !     Variable declarations.
@@ -298,7 +280,6 @@
 
           END IF
 
-
 !                                ============
           if (inum == num_emiss) exit SPCLOOP
 !                                ============
@@ -308,130 +289,6 @@
 !     ==============
       end do SPCLOOP
 !     ==============
-
-!     ----------------------
-!     GMI or GOCART aerosols
-!     ----------------------
-
-      if (emiss_timpyr == MONTHS_PER_YEAR) then
-        call GmiSplitDateTime  (nymd, idumyear, month, idumday)
-        it = month
-      else
-        it = 1
-      end if
-
-      if (emiss_aero_opt == 1) then                        ! GMI    aerosol emissions
-         emiss_aero(i1:i2,ju1:j2,1:naero) = emiss_aero_t(i1:i2,ju1:j2,1:naero,it)
-      elseif (emiss_aero_opt == 2) then                    ! GOCART aerosol emissions
-         emiss_aero(i1:i2,ju1:j2,1:5)     = emiss_aero_t(i1:i2,ju1:j2,1:5,it)
-      end if
-
-!     =====================================================
-      if ((emiss_aero_opt /= 0) .or. (emiss_dust_opt /= 0)) then
-!     =====================================================
-
-
-!       -------------------------------------------------------------------
-!       Add dust emission and other aerosols to some lowest vertical levels
-!       (emitted uniformly below the PBL).
-!       -------------------------------------------------------------------
-
-!       ----------------------------------------------------------------
-!       Add dust and aerosol emissions (kg/s) to concentrations; unit of
-!       aerosol and dust concentrations is kg/(kg air).
-!       ----------------------------------------------------------------
-
-!!!!! GOCART Source for dust
-        if (emiss_dust_opt == 2) then
-           call SourceDust (mcor, tdt, nymd, i1, i2, ju1, j2, k1, k2)
-           emiss_dust(i1:i2,ju1:j2,1:ndust) = srcEmissDust(i1:i2,ju1:j2,1:ndust)
-        end if
-!!!!! GOCART Source for sea salt
-        if (emiss_aero_opt == 2) then
-           call SourceSeaSalt (mcor, tdt, nymd, i1, i2, ju1, j2, k1, k2)
-           emiss_aero(i1:i2,ju1:j2,6:naero) = srcEmissSeaSalt(i1:i2,ju1:j2,1:4)
-        end if
-!!!!!
-
-        do ij = ju1, j2
-          do il = i1, i2
-
-            IKLOOP2: do ik = k1, k2
-
-              if ((za(il,ij,ik) > pbl(il,ij)) .and. (ik /= k1)) then
-!               ============
-                exit IKLOOP2
-!               ============
-              end if
-
-!             -----
-!             Dust.
-!             -----
-
-              if (emiss_dust_opt /= 0) then
-                do icx = 1, ndust
-
-                  ic = emiss_map_dust(icx)
-
-! Bigyani's correction
-!                 if (mass_pbl(il,ij) .le. 1.e-20)then
-!                    mass_pbl(il,ij) = 1.e-10
-!                 endif
-! End Bigyani's correction
-
-                  if (do_aerocom)then
-                     if ((ic /= INDMS) .and. (ik == k1)) then
-                       concentration(ic)%pArray3D(il,ij,ik) =  &
-     &                   concentration(ic)%pArray3D(il,ij,ik) +  &
-     &                     emiss_dust(il,ij,icx) * tdt/ mass(il,ij,ik)
-                     endif
-                  else
-                     concentration(ic)%pArray3D(il,ij,ik) =  &
-     &                 concentration(ic)%pArray3D(il,ij,ik) +  &
-     &                 emiss_dust(il,ij,icx) * tdt / mass_pbl(il,ij)
-                  end if
-
-                end do
-              end if
-
-!             ------------------------------
-!             Other aerosol (carbon & sslt).
-!             ------------------------------
-
-              if (emiss_aero_opt /= 0) then
-                do icx = 1, naero
-
-                  ic = emiss_map_aero(icx)
-
-                  if ((ic == IBOC) .or. (ic == IBBC)) then
-
-                    concentration(ic)%pArray3D(il,ij,ik) =  &
-     &                concentration(ic)%pArray3D(il,ij,ik) +  &
-     &                emiss_aero(il,ij,icx) * tdt / mass_pbl(il,ij)
-
-                  else if (((ic == INOC)   .or. (ic == IFOC)   .or.  &
-     &                      (ic == IFBC)   .or. (ic == ISSLT1) .or.  &
-     &                      (ic == ISSLT2) .or. (ic == ISSLT3) .or.  &
-     &                      (ic == ISSLT4)) .and.  &
-     &                     (ik == k1)) then
-
-                    concentration(ic)%pArray3D(il,ij,ik) =  &
-     &                concentration(ic)%pArray3D(il,ij,ik) +  &
-     &                emiss_aero(il,ij,icx) * tdt / mass(il,ij,ik)
-
-                  end if
-
-                end do
-              end if
-
-            end do IKLOOP2
-
-          end do
-        end do
-
-!     ======
-      end if
-!     ======
 
 
       return