+++ CHANGELOG.md
@@ -17,6 +17,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Deprecated
 
+
+### Added
+
+- Capability for AERO_PROVIDER=GMICHEM
+- Capability for AERO_PROVIDER=CARMA (uses GMICHEM right now)
+
+### Removed
+
+- Code that might have eventually run GOCART-like aerosols from within the GMI framework (note: doubtful this ever worked and would take significant effort to get working IMHO)
+
+
 ## [1.0.0] - 2023-01-18
 
 ### Added